(WIP) Config Jwt Token
